requirements.txt
numpy
pandas
opencv-python
matplotlib
scikit-learn
scikit-image
tensorflow
keras
streamlit
pillow
joblib
xgboost